+++ src/components/SummaryKeyComponents/SummaryKeyComponent.js
@@ -1,11 +1,12 @@
-import { AbstractComponent } from "./SummaryKeyComponents/AbstractComponent";
+import { SemScholarAbstractAndTldrComponent } from "./SemScholarAbstractAndTldrComponent";
+import { AbstractSummaryComponent } from "./AbstractSummaryComponent";
 
 /**
  * Component responsible for displaying the sidebar for summaries/key
  * @property {Object} components object that holds DOM elements that represent this component, as well as component's context
  * @property {HTMLElement} components.accordionItem accordion item
  * @property {HTMLElement} components.closeBtn button that closes sidepage
- * @property {AbstractComponent} abstractComponent abstract component instance
+ * @property {SemScholarAbstractAndTldrComponent} abstractAndTldrComponent
  * @property {AbstractSummaryComponent} abstractSummaryComponent abstract summary component instance
  */
 class SummaryKeyComponent {
@@ -20,7 +21,8 @@ class SummaryKeyComponent {
    * @constructor
    */
   constructor() {
-    this.tldrComponent = new AbstractComponent();
+    this.abstractAndTldrComponent = new SemScholarAbstractAndTldrComponent();
+    this.abstractSummaryComponent = new AbstractSummaryComponent();
     this.#registerEvents();
   }
 
@@ -48,7 +50,7 @@ class SummaryKeyComponent {
   createPageSummary = () => {};
 
   setPdf = (pdfDoc) => {
-    this.abstractComponent.setPdf(pdfDoc);
+    this.abstractAndTldrComponent.setPdf(pdfDoc);
   };
 }
