+++ src/templates/index.html
@@ -230,7 +230,7 @@
 
                     <button id="summary-abstract" class="accordion">ABSTRACT SUMMARY</button>
                     <div class="panel">
-                        <p id="abstractSummaryText">Abstract summary cannot be generated</p>
+                        <p id="abstract-summary-text">Abstract summary cannot be generated</p>
                     </div>
 
                     <button id="selected-text-summary" class="accordion">SELECTED TEXT SUMMARY</button>
